refactor(gui): move hooks to `hooks/GuiBuilder`
