+++ components/GuiBuilder/StageSettings/CompileSettings.tsx
@@ -1,5 +1,6 @@
 import { TagsInput, Textarea, TextInput } from "@components/Input";
-import { useSelectedStageConfig, useStoreState } from "@state/GuiBuilder/Hooks";
+import { useSelectedStageConfig } from "@hooks/GuiBuilder";
+import { useStoreState } from "@state/GuiBuilder/Hooks";
 import { Compile } from "@types";
 import { FocusEventHandler, memo } from "react";
 import { InfoTooltip } from "../Diagnostics";
