+++ components/GuiBuilder/StageSettings/StdioTestSettings/StdioTestCaseSettings.tsx
@@ -2,9 +2,10 @@ import Button from "@components/Button";
 import { MultiSelect, NumberInput, Select, SwitchGroup, Textarea, TextInput } from "@components/Input";
 import supportedStages, { valgrindDefaultConfig } from "@constants/GuiBuilder/supportedStages";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { useSelectedStageConfig } from "@hooks/GuiBuilder";
 import { clsx, Tooltip } from "@mantine/core";
 import { ControlledEditor, ControlledEditorProps } from "@monaco-editor/react";
-import { useSelectedStageConfig, useStoreActions, useStoreState } from "@state/GuiBuilder/Hooks";
+import { useStoreActions, useStoreState } from "@state/GuiBuilder/Hooks";
 import { StdioTest, TestCase } from "@types";
 import { getTestCasesLargestId } from "@utils/GuiBuilder/stageConfig";
 import cloneDeep from "lodash/cloneDeep";
