+++ components/GuiBuilder/StageSettings/StdioTestSettings/StdioTestStageSettings.tsx
@@ -1,5 +1,6 @@
 import { MultiSelect, TagsInput } from "@components/Input";
-import { useSelectedStageConfig, useStoreState } from "@state/GuiBuilder/Hooks";
+import { useSelectedStageConfig } from "@hooks/GuiBuilder";
+import { useStoreState } from "@state/GuiBuilder/Hooks";
 import { StdioTest } from "@types";
 import { FocusEventHandler, memo } from "react";
 import { InfoTooltip } from "../../Diagnostics";
