+++ tsconfig.json
@@ -28,6 +28,7 @@
       "@constants/*": ["constants/*"],
       "@contexts/*": ["contexts/*"],
       "@graphql/*": ["graphql/*"],
+      "@hooks/*": ["hooks/*"],
       "@layout": ["layout/"],
       "@lib/*": ["lib/*"],
       "@state/*": ["state/*"],
